requirements.txt
boto3==1.35.24
djangorestframework==3.15.2
pip-chill==1.0.3
polib==1.2.0
python-dotenv==1.0.1
